refactor: 보너스 볼 일치의 경우 0.5개 일치로 계산하기

+++ src/main/java/lotto/model/LottoWinner.java
@@ -10,7 +10,7 @@ public class LottoWinner {
         bonusLotto = bonus;
     }
 
-    public static int compare(Lotto lotto) {
+    public static float compare(Lotto lotto) {
         return winningLotto.countSameNumbers(lotto)
                 + bonusLotto.countSameNumber(lotto);
     }
